pyproject.toml
[project]
name = "wireframegen"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "flask-socketio>=5.3.0",
    "python-socketio>=5.0.0",
    "python-dotenv>=1.0.0",
]
